requirements.txt
# Core LangChain packages
langchain-core==0.3.55 # Core LangChain library
langchain-ollama==0.3.2 # Ollama integration for Langchain (local and remote)
langchain-chroma==0.2.3 # ChromaDB integration for vector storage

# Vector database
chromadb==0.6.3 # ChromaDB client for vector storage (remote)

# FastAPI framework and dependencies
fastapi==0.115.2 # Web framework for building APIs
uvicorn==0.30.6 # ASGI server for FastAPI
pydantic==2.11.3 # Data validation
pydantic-settings==2.3.0 # Settings management with Pydantic

# Logging
loguru==0.7.2 # Better logging for Python

# Utilities
python-dotenv==1.1.0 # For loading environment variables from .env files
numpy==1.26.4 # Used for array operations in testing
